moved processors into a separate package
